Start moving out the ingests stuff
