+++ archive/archive_api/src/tests/test_archive_api.py
@@ -3,31 +3,6 @@
 import json
 
 
-class TestReportIngestStatus:
-    """
-    Tests for the GET /ingests/<guid> endpoint.
-    """
-
-    def test_lookup_item(self, client):
-        lookup_id = "F423966E-A5E5-4D91-B321-88B90D1B5154"
-        resp = client.get(f"/storage/v1/ingests/{lookup_id}")
-        assert resp.status_code == 200
-        assert json.loads(resp.data) == {
-            "@context": "https://api.wellcomecollection.org/storage/v1/context.json",
-            "progress": lookup_id,
-        }
-
-    def test_lookup_missing_item_is_404(self, client):
-        lookup_id = "bad_status-404"
-        resp = client.get(f"/storage/v1/ingests/{lookup_id}")
-        assert resp.status_code == 404
-        assert (b"Invalid id: No ingest found for id=%r" % lookup_id) in resp.data
-
-    def test_post_against_lookup_endpoint_is_405(self, client, guid):
-        resp = client.post(f"/storage/v1/ingests/{guid}")
-        assert resp.status_code == 405
-
-
 class TestRequestNewIngest:
     """
     Tests for the POST /ingests endpoint.
